backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
phonenumbers==8.13.27
httpx==0.25.2 cle
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==20.7